requirements.txt
numpy >= 1.19.5
pandas >= 1.3.3
matplotlib >= 3.4.3
scikit-learn >= 0.24.2
scikit-optimize >= 0.8.1
scipy >= 1.7.1
linearpartition-unofficial>=0.3
gitlint
pre-commit
